packages/api/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
requests
PyYAML
# 以下库将在后续开发中添加
# supabase
# zep-python